refactor(dto): dto 패키지 제거
