+++ src/test/java/lotto/domain/WinningTicketTest.java
@@ -8,8 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import lotto.dto.PurchaseResult;
-
 public class WinningTicketTest {
 
     private LottoLine baseLottoLine;
